% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{meanRcpp}
\alias{meanRcpp}
\title{Mean value}
\usage{
meanRcpp(x)
}
\arguments{
\item{x}{a numericVector of samples}
}
\value{
mean value \code{m}
}
\description{
Calculate the mean value using Rcpp
}
\examples{
\dontrun{
x<-1:100
meanRcpp(x)
}
}
