% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGenoData.R
\name{rcbern}
\alias{rcbern}
\title{Generating High Dimension Correlated Binary Variables}
\usage{
rcbern(n, p, rho)
}
\arguments{
\item{n}{length of variables (numeric)}

\item{p}{success probility (numeric vector)}

\item{rho}{correlationship (numeric)}
}
\value{
Matrix of correlated bernoulli distribution variables \code{X}
}
\description{
A function to generate high dimension correlated bernoulli distribution variables with certain correlation and success probability.
}
\examples{
\dontrun{
p<-c(0.1,0.2,0.3)
rcbern(10,p,0.2)
}
}
