% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGenoData.R
\name{rcbern0}
\alias{rcbern0}
\title{Generating One Dimension Correlated Binary Variables}
\usage{
rcbern0(n, p, rho)
}
\arguments{
\item{n}{length of variables (numeric)}

\item{p}{success probility (numeric)}

\item{rho}{correlationship (numeric)}
}
\value{
Correlated bernoulli distribution variables \code{X}
}
\description{
A function to generate one dimension correlated bernoulli distribution variables with certain correlation and success probability.
}
\examples{
\dontrun{
rcbern0(10,0.5,0.2)
}
}
