% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGenoData.R
\name{rhoMaxEx}
\alias{rhoMaxEx}
\title{Upper bound for the correlation of correlated binary distribution}
\usage{
rhoMaxEx(p)
}
\arguments{
\item{p}{marginal frequencies (numeric vector)}
}
\value{
The upper bound for the pairwise correlation \code{rhoMax}
}
\description{
The upper bound for the pairwise correlation in exchange correlated binary distribution.
}
\examples{
\dontrun{
p=c(0.1,0.2,0.3)
rhoMaxEx(p)
}
}
