% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGenoData.R
\name{simGenoData}
\alias{simGenoData}
\title{Generating correlated genotypes}
\usage{
simGenoData(n, f, rho)
}
\arguments{
\item{n}{sample size (numeric)}

\item{f}{MAFs (numeric vector)}

\item{rho}{correlationship (numeric)}
}
\value{
Genotype Matrix \code{X}
}
\description{
A function to generate genotypes for correlated individuals at SNPs with certain MAFs.
}
\examples{
\dontrun{
f<-rep(0.2,10)
simGenoData(2,f,0.1)
}
}
