% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGenoData.R
\name{simGwasData}
\alias{simGwasData}
\title{Gwas simulation data}
\usage{
simGwasData(n0, n1, f0, f1, rho)
}
\arguments{
\item{n0}{case size (numeric)}

\item{n1}{control size (numeric)}

\item{f0}{MAFs of cases (numeric vector)}

\item{f1}{MAFs of controls (numeric vector)}

\item{rho}{correlationship between all individuals (numeric)}
}
\value{
Genotype Matrix \code{X}

label of case or control \code{y}
}
\description{
A function to generate genotypes when both cases and controls are correlated.
}
\examples{
\dontrun{
f0<-f1<-rep(0.2,100)
simGwasData(10,10,f0,f1,0.1)
}
}
