% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGenoData.R
\name{simGwasDataII}
\alias{simGwasDataII}
\title{Gwas simulation data II}
\usage{
simGwasDataII(n0, n1, f0, f1, rho0, rho1)
}
\arguments{
\item{n0}{case size (numeric)}

\item{n1}{control size (numeric)}

\item{f0}{MAFs of cases (numeric vector)}

\item{f1}{MAFs of controls (numeric vector)}

\item{rho0}{correlationship between cases (numeric)}

\item{rho1}{correlationship between controls (numeric)}
}
\value{
Genotype Matrix \code{X}

label of case or control \code{y}
}
\description{
A function to generate genotypes when different correlations are used among cases and controls, but no correlations between a case and a control.
}
\examples{
\dontrun{
f0<-f1<-rep(0.3,100)
simGwasDataII(10,10,f0,f1,0.1,0.2)
}
}
